+++ zerver/models.py
@@ -4543,6 +4543,9 @@ class ScheduledMessage(models.Model):
     delivered_message = models.ForeignKey(Message, null=True, on_delete=CASCADE)
     has_attachment = models.BooleanField(default=False, db_index=True)
 
+    # Only used for REMIND delivery_type messages.
+    reminder_target_message_id = models.IntegerField(null=True)
+
     # Metadata for messages that failed to send when their scheduled
     # moment arrived.
     failed = models.BooleanField(default=False)